+++ greentest/greentest.py
@@ -71,12 +71,19 @@ PLATFORM_SPECIFIC_SUFFIXES = ['2', '279', '3']
 if sys.platform.startswith('win'):
     PLATFORM_SPECIFIC_SUFFIXES.append('posix')
 
+PY2 = None
+PY3 = None
+PY34 = None
+
 NON_APPLICABLE_SUFFIXES = []
 if sys.version_info[0] == 3:
     # Python 3
     NON_APPLICABLE_SUFFIXES.extend(('2', '279'))
     PY2 = False
     PY3 = True
+    if sys.version_info[1] >= 4:
+        PY34 = True
+
 if sys.version_info[0] == 2:
     # Any python 2
     PY3 = False
@@ -155,6 +162,11 @@ def wrap_timeout(timeout, method):
             return method(self, *args, **kwargs)
 
     return wrapped
+
+
+def ignores_leakcheck(func):
+    func.ignore_leakcheck = True
+    return func
 
 
 def wrap_refcount(method):